pyproject.toml
[project]
name = "vtc-link"
version = "1.0.0"
description = "Hospital vital data interface proxy - Normalizes patient vital signs from various hospital systems and forwards to backend analysis servers"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "VTC-Link Contributors"}
]
keywords = ["healthcare", "vital-signs", "hospital", "integration", "fastapi", "hl7", "medical"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Framework :: FastAPI",
]
requires-python = ">=3.12,<3.13"

dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "pyyaml>=6.0.1",
  "duckdb>=0.10.0",
  "httpx>=0.27.0",
  "apscheduler>=3.10.4",
  "jinja2>=3.1.3",
  "python-multipart>=0.0.9",
  "oracledb>=2.2.0",
  "pyodbc>=5.1.0",
]


[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.uv]
managed = true

[tool.setuptools]
packages = ["app"]
include-package-data = true

[tool.setuptools.package-data]
app = ["templates/**/*.html", "static/*.css"]

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"
